Added const a few more places

+++ src/vmem/vmem_client.c
@@ -67,7 +67,7 @@ void vmem_download(int node, int timeout, uint32_t address, uint32_t length, cha
 
 	uint32_t time_total = csp_get_ms() - time_begin;
 
-	printf("  Downloaded %u bytes in %.03f s at %u Bps\n", length, time_total / 1000.0, (unsigned int) (length / ((float)time_total / 1000.0)) );
+	printf("  Downloaded %u bytes in %.03f s at %u Bps\n", (unsigned int) length, time_total / 1000.0, (unsigned int) (length / ((float)time_total / 1000.0)) );
 
 
 }
@@ -131,7 +131,7 @@ void vmem_upload(int node, int timeout, uint32_t address, char * datain, uint32_
 
 	uint32_t time_total = csp_get_ms() - time_begin;
 
-	printf("  Uploaded %u bytes in %.03f s at %u Bps\n", length, time_total / 1000.0, (unsigned int) (length / ((float)time_total / 1000.0)) );
+	printf("  Uploaded %u bytes in %.03f s at %u Bps\n", (unsigned int) length, time_total / 1000.0, (unsigned int) (length / ((float)time_total / 1000.0)) );
 
 }
